+++ jax/_src/pallas/pallas_call.py
@@ -1549,7 +1549,7 @@ def pallas_call(
       {file}:{line}`.
     compiler_params: Optional compiler parameters. The value should either be a
       backend-specific dataclass
-      (:class:`jax.experimental.pallas.tpu.TPUCompilerParams`,
+      (:class:`jax.experimental.pallas.tpu.CompilerParams`,
       :class:`jax.experimental.pallas.triton.CompilerParams`,
       :class:`jax.experimental.pallas.mosaic_gpu.CompilerParams`) or a dict
       mapping backend name to the corresponding platform-specific dataclass.