+++ tests/pallas/tpu_pallas_interpret_distributed_test.py
@@ -92,7 +92,7 @@ class InterpretDistributedTest(jtu.JaxTestCase):
     out_shape = jax.ShapeDtypeStruct((8, 128), jnp.float32)
     grid_spec = pltpu.PrefetchScalarGridSpec(
         num_scalar_prefetch=0,
-        # TPUMemorySpace.ANY will (usually) place the tensor in HBM.
+        # MemorySpace.ANY will (usually) place the tensor in HBM.
         in_specs=[
             pl.BlockSpec(memory_space=pltpu.ANY),
         ],
@@ -106,7 +106,7 @@ class InterpretDistributedTest(jtu.JaxTestCase):
         right_permute_kernel,
         out_shape=out_shape,
         grid_spec=grid_spec,
-        compiler_params=pltpu.TPUCompilerParams(collective_id=13),
+        compiler_params=pltpu.CompilerParams(collective_id=13),
         interpret=mosaic_interpret.TPUInterpretParams(
             dma_execution_mode=dma_execution_mode, detect_races=detect_races),
     )
@@ -206,7 +206,7 @@ class InterpretDistributedTest(jtu.JaxTestCase):
     grid_spec = pltpu.PrefetchScalarGridSpec(
       num_scalar_prefetch=0,
       in_specs=[
-        # TPUMemorySpace.ANY will (usually) place the tensor in HBM.
+        # MemorySpace.ANY will (usually) place the tensor in HBM.
         pl.BlockSpec(memory_space=pltpu.ANY),
       ],
       out_specs=pl.BlockSpec(memory_space=pltpu.ANY),
@@ -230,7 +230,7 @@ class InterpretDistributedTest(jtu.JaxTestCase):
       grid_spec=grid_spec,
       interpret=mosaic_interpret.TPUInterpretParams(
           dma_execution_mode=dma_execution_mode, detect_races=detect_races),
-      compiler_params=pltpu.TPUCompilerParams(collective_id=0),
+      compiler_params=pltpu.CompilerParams(collective_id=0),
     )
 
     # Wrap the kernel within a shard_map to call.
@@ -390,7 +390,7 @@ class InterpretDistributedTest(jtu.JaxTestCase):
       grid_spec=grid_spec,
       interpret=mosaic_interpret.TPUInterpretParams(
           dma_execution_mode=dma_execution_mode, detect_races=detect_races),
-      compiler_params=pltpu.TPUCompilerParams(collective_id=0),
+      compiler_params=pltpu.CompilerParams(collective_id=0),
     )
 
     pallas_result = jax.jit(
@@ -674,7 +674,7 @@ class InterpretDistributedTest(jtu.JaxTestCase):
         grid_spec=grid_spec,
         interpret=mosaic_interpret.TPUInterpretParams(
             dma_execution_mode=dma_execution_mode, detect_races=True),
-        compiler_params=pltpu.TPUCompilerParams(collective_id=7),
+        compiler_params=pltpu.CompilerParams(collective_id=7),
       )(input_arr)[0]
 
     pallas_result = jax.jit(
@@ -978,7 +978,7 @@ class InterpretDistributedTest(jtu.JaxTestCase):
         grid_spec=grid_spec,
         interpret=mosaic_interpret.TPUInterpretParams(
             dma_execution_mode=dma_execution_mode, detect_races=detect_races),
-        compiler_params=pltpu.TPUCompilerParams(collective_id=19),
+        compiler_params=pltpu.CompilerParams(collective_id=19),
       )(input_arr)[0]
 
     pallas_result = jax.jit(